+++ common/src/main/java/net/mehvahdjukaar/moonlight/api/misc/StreamCodecMapRegistry.java
@@ -1,10 +1,7 @@
 package net.mehvahdjukaar.moonlight.api.misc;
 
-import com.mojang.serialization.MapCodec;
-import net.mehvahdjukaar.moonlight.api.set.BlocksColorAPI;
 import net.minecraft.network.RegistryFriendlyByteBuf;
 import net.minecraft.network.codec.StreamCodec;
-import net.minecraft.resources.ResourceLocation;
 
 // simple registry of stream codecs
 public class StreamCodecMapRegistry<T> extends MapRegistry<StreamCodec<? super RegistryFriendlyByteBuf, ? extends T>> {