+++ common/src/main/java/net/mehvahdjukaar/moonlight/api/misc/WorldSavedDataType.java
@@ -4,6 +4,7 @@ package net.mehvahdjukaar.moonlight.api.misc;
 import com.mojang.serialization.Codec;
 import net.mehvahdjukaar.moonlight.api.MoonlightRegistry;
 import net.mehvahdjukaar.moonlight.api.platform.PlatHelper;
+import net.mehvahdjukaar.moonlight.core.set.BlocksColorInternal;
 import net.minecraft.core.HolderLookup;
 import net.minecraft.nbt.CompoundTag;
 import net.minecraft.nbt.NbtOps;
@@ -14,7 +15,6 @@ import net.minecraft.resources.ResourceLocation;
 import net.minecraft.server.level.ServerLevel;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.saveddata.SavedData;
-import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
 import java.util.function.Function;
@@ -22,28 +22,51 @@ import java.util.function.Function;
 //Basically a helper that manages syncing on world join and stores codecs cleanly
 //per world, automatically synced when stream codec is not null. not per game. saved onto overworld level
 public final class WorldSavedDataType<D extends WorldSavedData> {
-
-    public static final Codec<WorldSavedDataType<? extends WorldSavedData>> CODEC =
-            MoonlightRegistry.WORLD_SAVED_DATA_TYPE_REGISTRY.byNameCodec();
     public static final StreamCodec<RegistryFriendlyByteBuf, WorldSavedDataType<? extends WorldSavedData>> STREAM_CODEC =
             ByteBufCodecs.registry(MoonlightRegistry.WORLD_SAVED_DATA_TYPE_REGISTRY.key());
+    public static final Codec<WorldSavedDataType<? extends WorldSavedData>> CODEC =
+            MoonlightRegistry.WORLD_SAVED_DATA_TYPE_REGISTRY.byNameCodec();
 
+    //TODO: 1.22 make map codec here instead
 
     private final Codec<D> codec;
     @Nullable
-    private final StreamCodec<RegistryFriendlyByteBuf, D> streamCodec;
+    private final StreamCodec<? super RegistryFriendlyByteBuf, D> streamCodec;
     private final SavedData.Factory<D> factory;
     private final String name;
+    private final Scope scope;
 
 
     private D clientInstance = null;
 
-    public WorldSavedDataType(ResourceLocation id, Function<ServerLevel, D> constructor, Codec<D> codec, @Nullable StreamCodec<RegistryFriendlyByteBuf, D> streamCodec) {
+    public enum Scope {
+        SINGLE_OVERWORLD,
+        PER_LEVEL;
+
+        private ServerLevel getTargetLevel(ServerLevel sl) {
+            ServerLevel targetLevel;
+            if (this == Scope.PER_LEVEL) {
+                targetLevel = sl;
+            } else {
+                targetLevel = sl.getServer().overworld();
+            }
+            return targetLevel;
+        }
+    }
+
+    public WorldSavedDataType(ResourceLocation id, Function<ServerLevel, D> overworldToDataConstructor, Codec<D> codec,
+                              @Nullable StreamCodec<? super RegistryFriendlyByteBuf, D> streamCodec) {
+        this(id, overworldToDataConstructor, codec, streamCodec, Scope.SINGLE_OVERWORLD);
+    }
+
+    public WorldSavedDataType(ResourceLocation id, Function<ServerLevel, D> overworldToDataConstructor, Codec<D> codec,
+                              @Nullable StreamCodec<? super RegistryFriendlyByteBuf, D> streamCodec, Scope scope) {
         this.codec = codec;
         this.streamCodec = streamCodec;
         this.name = id.toDebugFileName();
+        this.scope = scope;
 
-        this.factory = new SavedData.Factory<>(() -> constructor.apply(
+        this.factory = new SavedData.Factory<>(() -> overworldToDataConstructor.apply(
                 PlatHelper.getCurrentServer().overworld()),
                 this::load, null);
     }
@@ -56,16 +79,18 @@ public final class WorldSavedDataType<D extends WorldSavedData> {
         }
 
         if (level instanceof ServerLevel server) {
-            return server.getServer().overworld().getDataStorage()
-                    .computeIfAbsent(factory, name);
+            ServerLevel targetLevel = scope.getTargetLevel(server);
+            return targetLevel.getDataStorage().computeIfAbsent(factory, name);
         } else {
             return clientInstance;
         }
     }
 
+
     public void setData(Level level, D data) {
         if (level instanceof ServerLevel server) {
-            server.getServer().overworld().getDataStorage().set(name, data);
+            ServerLevel targetLevel = scope.getTargetLevel(server);
+            targetLevel.getDataStorage().set(name, data);
         } else {
             this.clientInstance = data;
         }
@@ -79,12 +104,13 @@ public final class WorldSavedDataType<D extends WorldSavedData> {
         return dataResult.getOrThrow().getFirst();
     }
 
+    //TODO: make map codec
     public Codec<D> getCodec() {
         return codec;
     }
 
     @Nullable
-    public StreamCodec<RegistryFriendlyByteBuf, D> getStreamCodec() {
+    public StreamCodec<? super RegistryFriendlyByteBuf, D> getStreamCodec() {
         return streamCodec;
     }